+++ src/RcppExports.cpp
@@ -51,6 +51,22 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// calc_freq_local_pwm_cpp
+Rcpp::List calc_freq_local_pwm_cpp(const Rcpp::List& freqs, const Rcpp::NumericMatrix& pwm, const Rcpp::NumericMatrix& pwm_rc, const Rcpp::IntegerVector& motif_lengths, const bool multiply, const bool bidirect);
+RcppExport SEXP _prego_calc_freq_local_pwm_cpp(SEXP freqsSEXP, SEXP pwmSEXP, SEXP pwm_rcSEXP, SEXP motif_lengthsSEXP, SEXP multiplySEXP, SEXP bidirectSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Rcpp::List& >::type freqs(freqsSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::NumericMatrix& >::type pwm(pwmSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::NumericMatrix& >::type pwm_rc(pwm_rcSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::IntegerVector& >::type motif_lengths(motif_lengthsSEXP);
+    Rcpp::traits::input_parameter< const bool >::type multiply(multiplySEXP);
+    Rcpp::traits::input_parameter< const bool >::type bidirect(bidirectSEXP);
+    rcpp_result_gen = Rcpp::wrap(calc_freq_local_pwm_cpp(freqs, pwm, pwm_rc, motif_lengths, multiply, bidirect));
+    return rcpp_result_gen;
+END_RCPP
+}
 // pssm_dataset_cor_parallel
 NumericMatrix pssm_dataset_cor_parallel(const List& pssm_list, const std::string& method, double prior);
 RcppExport SEXP _prego_pssm_dataset_cor_parallel(SEXP pssm_listSEXP, SEXP methodSEXP, SEXP priorSEXP) {
@@ -260,6 +276,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_prego_n_nuc_distribution", (DL_FUNC) &_prego_n_nuc_distribution, 3},
     {"_prego_calc_sequences_dinuc_cpp", (DL_FUNC) &_prego_calc_sequences_dinuc_cpp, 1},
     {"_prego_kmer_matrix_cpp", (DL_FUNC) &_prego_kmer_matrix_cpp, 7},
+    {"_prego_calc_freq_local_pwm_cpp", (DL_FUNC) &_prego_calc_freq_local_pwm_cpp, 6},
     {"_prego_pssm_dataset_cor_parallel", (DL_FUNC) &_prego_pssm_dataset_cor_parallel, 3},
     {"_prego_calc_seq_pwm_parallel_cpp", (DL_FUNC) &_prego_calc_seq_pwm_parallel_cpp, 8},
     {"_prego_get_consensus_cpp", (DL_FUNC) &_prego_get_consensus_cpp, 3},